we don't actually need thread safety in the Searcher bitset, just the ability to grow itself
extract JavaUtilBitSet -> GrowableBitSet to handle this use case

this makes concurrent build about 8% faster (~162s on Texmex benchmark to ~150)
